+++ Changelog.md
@@ -8,6 +8,7 @@ Compiler Features:
  * Optimizer: Add rule to simplify SUB(~0, X) to NOT(X).
  * Commandline Interface: Experimental parser error recovery via the ``--error-recovery`` commandline switch.
  * Yul Optimizer: Make the optimizer work for all dialects of Yul including eWasm.
+ * Metadata: Update the swarm hash, changes ``bzzr0`` to ``bzzr1`` and urls to use ``bzz-raw://``.